requirements.txt
fastapi
uvicorn
streamlit
textblob
python-dotenv
requests
pydantic
